# openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Dorky Artifact Service
  description: Language-agnostic HTTP service for artifact storage, streaming, metadata, idempotency, and versioning
  version: 0.2.0
servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /health:
    get:
      summary: Health check
      operationId: health
      tags:
        - service
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  timestamp:
                    type: string
                    format: date-time

  /artifacts:
    post:
      summary: Upload an artifact
      operationId: uploadArtifact
      tags:
        - artifacts
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload
                metadata:
                  type: string
                  description: JSON metadata as string
                idempotency_key:
                  type: string
                  description: Unique key to prevent duplicate uploads
              required:
                - file
      responses:
        '201':
          description: Artifact created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artifact'
        '400':
          description: Bad request (missing file, invalid metadata)
        '409':
          description: Conflict (artifact already exists with same idempotency key)

  /artifacts/{id}:
    get:
      summary: Download an artifact
      operationId: downloadArtifact
      tags:
        - artifacts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Artifact ID
        - name: version
          in: query
          required: false
          schema:
            type: string
          description: Specific version to download (optional)
      responses:
        '200':
          description: Artifact stream
          content:
            application/octet-stream: {}
        '404':
          description: Artifact not found
        '416':
          description: Range not satisfiable

    delete:
      summary: Delete an artifact
      operationId: deleteArtifact
      tags:
        - artifacts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Artifact deleted successfully
        '404':
          description: Artifact not found

  /artifacts/{id}/metadata:
    get:
      summary: Get artifact metadata
      operationId: getMetadata
      tags:
        - metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Artifact metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metadata'
        '404':
          description: Artifact not found

    patch:
      summary: Update artifact metadata
      operationId: updateMetadata
      tags:
        - metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Metadata updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metadata'
        '404':
          description: Artifact not found

  /artifacts/{id}/versions:
    get:
      summary: List all versions of an artifact
      operationId: listVersions
      tags:
        - versioning
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Version'
        '404':
          description: Artifact not found

components:
  schemas:
    Artifact:
      type: object
      required:
        - id
        - url
      properties:
        id:
          type: string
          example: "1a2b3c"
        url:
          type: string
          example: "/artifacts/1a2b3c"
        metadata:
          $ref: '#/components/schemas/Metadata'
        version:
          type: string
          example: "1"

    Metadata:
      type: object
      required:
        - content_hash
        - file_size
        - last_updated
      properties:
        content_hash:
          type: string
          description: SHA256 hash of file content
          example: "abc123def456..."
        file_size:
          type: integer
          description: File size in bytes
          example: 1024
        compressed_size:
          type: integer
          description: Compressed file size (if applicable)
        compression_type:
          type: string
          description: Compression algorithm used
          enum: [gzip, lz4, brotli, null]
        last_updated:
          type: string
          format: date-time
          description: Last update timestamp (UTC ISO 8601)
        uploaded_by:
          type: string
          description: User/identifier who uploaded
        custom:
          type: object
          description: User-defined metadata
          additionalProperties: true

    Version:
      type: object
      required:
        - version_id
        - created_at
        - metadata
      properties:
        version_id:
          type: string
          example: "1"
        created_at:
          type: string
          format: date-time
        metadata:
          $ref: '#/components/schemas/Metadata'